requirements.txt

openai
python-dotenv
pandas

streamlit

pytest
